requirements.txt
mcp[cli]
openai-agents
python-dotenv
httpx
